Modify template to use the new format
