+++ at_app_flutter/lib/src/templates/app/lib/home_screen.dart
@@ -1,0 +1,28 @@
+import 'package:at_client_mobile/at_client_mobile.dart';
+import 'package:flutter/material.dart';
+
+// * Once the onboarding process is completed you will be taken to this screen
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    // * Getting the AtClientManager instance to use below
+    AtClientManager atClientManager = AtClientManager.getInstance();
+
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('What\'s my current @sign?'),
+      ),
+      body: Center(
+        child: Column(children: [
+          const Text('Successfully onboarded and navigated to FirstAppScreen'),
+
+          // * Use the AtClientManager instance to get the AtClient
+          // * Then use the AtClient to get the current @sign
+          Text('Current @sign: ${atClientManager.atClient.getCurrentAtSign()}')
+        ]),
+      ),
+    );
+  }
+}
